将ChunkedList类的计数和索引类型从int32_t改为int64_t以支持更大数据量

+++ include/base/container/ChunkedList.h
@@ -50,7 +50,7 @@ namespace base
 		///
 		/// @param size
 		///
-		void Reserve(int32_t size)
+		void Reserve(int64_t size)
 		{
 			if (size < 0)
 			{
@@ -80,9 +80,9 @@ namespace base
 		/// @param index
 		/// @param item
 		///
-		virtual void Insert(int32_t const index, ItemType const &item) override
+		virtual void Insert(int64_t const index, ItemType const &item) override
 		{
-			if (index < 0 || index > static_cast<int32_t>(_deque.size()))
+			if (index < 0 || index > static_cast<int64_t>(_deque.size()))
 			{
 				throw std::out_of_range{"索引超出范围"};
 			}
@@ -127,9 +127,9 @@ namespace base
 		///
 		/// @param index
 		///
-		virtual void RemoveAt(int32_t const index) override
+		virtual void RemoveAt(int64_t const index) override
 		{
-			if (index < 0 || index >= static_cast<int32_t>(_deque.size()))
+			if (index < 0 || index >= static_cast<int64_t>(_deque.size()))
 			{
 				throw std::out_of_range{"索引超出范围"};
 			}
@@ -188,7 +188,7 @@ namespace base
 		///
 		/// @return
 		///
-		virtual int32_t IndexOf(ItemType const &item) const override
+		virtual int64_t IndexOf(ItemType const &item) const override
 		{
 			auto it = std::find_if(_deque.begin(),
 								   _deque.end(),
@@ -199,7 +199,7 @@ namespace base
 
 			if (it != _deque.end())
 			{
-				return static_cast<int32_t>(std::distance(_deque.begin(), it)); // 返回元素的索引
+				return static_cast<int64_t>(std::distance(_deque.begin(), it)); // 返回元素的索引
 			}
 
 			// 如果没有找到元素，返回 -1
@@ -232,9 +232,9 @@ namespace base
 		///
 		/// @return
 		///
-		virtual int32_t Count() const override
+		virtual int64_t Count() const override
 		{
-			return static_cast<int32_t>(_deque.size());
+			return static_cast<int64_t>(_deque.size());
 		}
 
 		/* #region Sort */
@@ -322,7 +322,7 @@ namespace base
 		///
 		/// @return
 		///
-		virtual ItemType &operator[](int32_t const index) override
+		virtual ItemType &operator[](int64_t const index) override
 		{
 			return _deque[index];
 		}
@@ -343,7 +343,7 @@ namespace base
 		///
 		/// @return
 		///
-		virtual ItemType const &operator[](int32_t const index) const override
+		virtual ItemType const &operator[](int64_t const index) const override
 		{
 			return _deque[index];
 		}
